+++ DevelopGitHub/GitHubExtension/GitHubExtension.Security.WebApi.Library/Controllers/RolesController.cs
@@ -2,7 +2,7 @@
 using System.Web.Http;
 using System.Web.Http.Cors;
 using GitHubExtension.Security.DAL.Interfaces;
-using GitHubExtension.Security.WebApi.Library.Converters;
+using GitHubExtension.Security.WebApi.Library.Mappers;
 
 namespace GitHubExtension.Security.WebApi.Library.Controllers
 {
