+++ DevelopGitHub/GitHubExtension/GitHubExtension.Security.WebApi.Library/Mappers/RepositoryMapper.cs
@@ -1,9 +1,9 @@
 ﻿using GitHubExtension.Models.CommunicationModels;
 using GitHubExtension.Security.StorageModels.Identity;
 
-namespace GitHubExtension.Security.WebApi.Library.Converters
+namespace GitHubExtension.Security.WebApi.Library.Mappers
 {
-    public static class RepositoryConverter
+    public static class RepositoryMapper
     {
         public static Repository ToEntity(this RepositoryDto repository)
         {
